Kind of psuedo-code refactor towards a model I like
