--- src/mod.rs
@@ -1,2 +1,0 @@
-pub mod types;
-pub mod error;
