+++ src/types/accuracy_data.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Debug, PartialEq)]
+pub struct AccuracyData {
+    pub overall: Fraction,
+    pub per_key: HashMap<KeyEvent, Fraction>,
+}
