+++ src/types/fraction.rs
@@ -1,0 +1,38 @@
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+pub struct Fraction {
+    pub numerator: usize,
+    pub denominator: usize,
+}
+
+impl Fraction {
+    pub const fn new(numerator: usize, denominator: usize) -> Self {
+        Self {
+            numerator,
+            denominator,
+        }
+    }
+}
+
+impl From<Fraction> for f64 {
+    fn from(f: Fraction) -> Self {
+        f.numerator as f64 / f.denominator as f64
+    }
+}
+
+impl cmp::Ord for Fraction {
+    fn cmp(&self, other: &Self) -> cmp::Ordering {
+        f64::from(*self).partial_cmp(&f64::from(*other)).unwrap()
+    }
+}
+
+impl PartialOrd for Fraction {
+    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl fmt::Display for Fraction {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}/{}", self.numerator, self.denominator)
+    }
+}
