+++ src/types/test/handler.rs
@@ -1,0 +1,103 @@
+impl Test {
+    pub fn handle_key(&mut self, key: KeyEvent) {
+        if key.kind != KeyEventKind::Press {
+            return;
+        }
+
+        let word = &mut self.words[self.current_word];
+        match key.code {
+            KeyCode::Char(' ') | KeyCode::Enter => {
+                if word.text.chars().nth(word.progress.len()) == Some(' ') {
+                    word.progress.push(' ');
+                    word.events.push(TestEvent {
+                        time: Instant::now(),
+                        correct: Some(true),
+                        key,
+                    })
+                } else if !word.progress.is_empty() || word.text.is_empty() {
+                    let correct = word.text == word.progress;
+                    if self.sudden_death_enabled && !correct {
+                        self.reset();
+                    } else {
+                        word.events.push(TestEvent {
+                            time: Instant::now(),
+                            correct: Some(correct),
+                            key,
+                        });
+                        self.next_word();
+                    }
+                }
+            }
+            KeyCode::Backspace => {
+                if word.progress.is_empty() && self.backtracking_enabled && self.backspace_enabled {
+                    self.last_word();
+                } else if self.backspace_enabled {
+                    word.events.push(TestEvent {
+                        time: Instant::now(),
+                        correct: Some(!word.text.starts_with(&word.progress[..])),
+                        key,
+                    });
+                    word.progress.pop();
+                }
+            }
+            // CTRL-BackSpace and CTRL-W
+            KeyCode::Char('h') | KeyCode::Char('w')
+                if key.modifiers.contains(KeyModifiers::CONTROL) =>
+            {
+                if self.words[self.current_word].progress.is_empty() {
+                    self.last_word();
+                }
+
+                let word = &mut self.words[self.current_word];
+
+                word.events.push(TestEvent {
+                    time: Instant::now(),
+                    correct: None,
+                    key,
+                });
+                word.progress.clear();
+            }
+            KeyCode::Char(c) => {
+                word.progress.push(c);
+                let correct = word.text.starts_with(&word.progress[..]);
+                if self.sudden_death_enabled && !correct {
+                    self.reset();
+                } else {
+                    word.events.push(TestEvent {
+                        time: Instant::now(),
+                        correct: Some(correct),
+                        key,
+                    });
+                    if word.progress == word.text && self.current_word == self.words.len() - 1 {
+                        self.complete = true;
+                        self.current_word = 0;
+                    }
+                }
+            }
+            _ => {}
+        };
+    }
+    fn last_word(&mut self) {
+        if self.current_word != 0 {
+            self.current_word -= 1;
+        }
+    }
+
+    fn next_word(&mut self) {
+        if self.current_word == self.words.len() - 1 {
+            self.complete = true;
+            self.current_word = 0;
+        } else {
+            self.current_word += 1;
+        }
+    }
+
+    fn reset(&mut self) {
+        self.words.iter_mut().for_each(|word: &mut TestWord| {
+            word.progress.clear();
+            word.events.clear();
+        });
+        self.current_word = 0;
+        self.complete = false;
+    }
+}
