+++ src/types/test_word.rs
@@ -1,0 +1,22 @@
+#[derive(Debug, Clone, PartialEq)]
+pub struct TestWord {
+    pub text: String,
+    pub progress: String,
+    pub events: Vec<TestEvent>,
+}
+
+impl From<String> for TestWord {
+    fn from(string: String) -> Self {
+        TestWord {
+            text: string,
+            progress: String::new(),
+            events: Vec::new(),
+        }
+    }
+}
+
+impl From<&str> for TestWord {
+    fn from(string: &str) -> Self {
+        Self::from(string.to_string())
+    }
+}
